refactor: split fuzz.rs god-file into modules
