+++ homeboy.json
@@ -662,13 +662,13 @@
         "dead_code::src/core/agent_task_dispatch_service.rs::UnreferencedExport",
         "dead_code::src/core/daemon/control.rs::UnreferencedExport",
         "dead_code::src/core/deterministic_loop.rs::UnreferencedExport",
-        "dead_code::src/core/fuzz.rs::UnreferencedExport",
+        "dead_code::src/core/fuzz/types.rs::UnreferencedExport",
         "dead_code::src/core/lab_routing.rs::UnreferencedExport",
         "docs::docs/commands/runtime.md::BrokenDocReference",
         "duplication::src/core/artifact_contract.rs::DuplicateFunction",
         "duplication::src/core/extension/runner.rs::DuplicateFunction",
         "duplication::src/core/extension/test/run.rs::DuplicateFunction",
-        "duplication::src/core/fuzz.rs::DuplicateFunction",
+        "duplication::src/core/fuzz/types.rs::DuplicateFunction",
         "field_patterns::src/commands/agent_task/args/controller.rs::RepeatedFieldPattern",
         "field_patterns::src/commands/fuzz.rs::RepeatedFieldPattern",
         "field_patterns::src/commands/report/performance_digest.rs::RepeatedFieldPattern",
@@ -690,7 +690,7 @@
         "field_patterns::src/core/extension/bench/report.rs::RepeatedFieldPattern",
         "field_patterns::src/core/extension/bench/responsiveness.rs::RepeatedFieldPattern",
         "field_patterns::src/core/extension/trace/report.rs::RepeatedFieldPattern",
-        "field_patterns::src/core/fuzz.rs::RepeatedFieldPattern",
+        "field_patterns::src/core/fuzz/types.rs::RepeatedFieldPattern",
         "field_patterns::src/core/git/github_types.rs::RepeatedFieldPattern",
         "field_patterns::src/core/git/pr_land.rs::RepeatedFieldPattern",
         "field_patterns::src/core/git/pr_policy.rs::RepeatedFieldPattern",
@@ -736,7 +736,7 @@
         "intra-method-duplication::src/core/runner/worker/run.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/upgrade/runners.rs::IntraMethodDuplicate",
         "near-duplication::src/core/artifact_contract.rs::NearDuplicate",
-        "near-duplication::src/core/fuzz.rs::NearDuplicate",
+        "near-duplication::src/core/fuzz/types.rs::NearDuplicate",
         "parallel-implementation::src/core/agent_task_prompts.rs::ParallelImplementation",
         "parallel-implementation::src/core/extension/trace/attach.rs::ParallelImplementation",
         "structural::src/cli_surface.rs::HighItemCount",
@@ -786,8 +786,6 @@
         "structural::src/core/extension/execution.rs::HighItemCount",
         "structural::src/core/extension/grammar.rs::GodFile",
         "structural::src/core/extension/trace/report.rs::HighItemCount",
-        "structural::src/core/fuzz.rs::GodFile",
-        "structural::src/core/fuzz.rs::HighItemCount",
         "structural::src/core/git/github.rs::GodFile",
         "structural::src/core/git/github.rs::HighItemCount",
         "structural::src/core/git/github_types.rs::HighItemCount",
