+++ src/core/fuzz/tests/mod.rs
@@ -1,0 +1,6 @@
+//! Tests for the product-neutral fuzz contracts.
+
+mod campaign_tests;
+mod contract_tests;
+mod parse_tests;
+mod surface_tests;
